员工详情页面修改

+++ src/main/java/com/demo/service/UserService.java
@@ -4,5 +4,5 @@ import org.springframework.stereotype.Service;
 
 @Service
 public interface UserService {
-     boolean findUser(Integer id, String password);
+     boolean findUser(Long id, String password);
 }
